pyproject.toml
[project]
name = "brilliant-minds"
version = "0.1.0"
description = "AI assistant to simplify documents for people with dyslexia and ADHD"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0rc4",
    "agent-framework-azure-ai>=1.0.0rc4",
    "azure-search-documents>=11.7.0b2",
    "azure-storage-blob>=12.29.0b1",
    "bcrypt>=5.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.3.0",
]

[tool.hatch.package]
src = "src"

[tool.uv]
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'"
]
